requirements.txt
numpy>=1.26,<3
pandas>=2.2,<3
scipy>=1.11,<2
scikit-learn>=1.4,<2
xgboost>=2.0,<3
cvxpy>=1.4,<2
PyYAML>=6.0,<7
python-dotenv>=1.0,<2
ib-insync>=0.9.86,<1
yfinance>=0.2.50,<1
pyarrow>=16,<20
matplotlib>=3.8,<4
pytest>=8.0,<9
